VS 2013 emits a warning when two source files have the same name; but it also does this when one of the files is excluded from the build, which I think is a bug.
Adding the IgnoreWarnCompileDuplicatedFilename key will
suppress this warning; it does so also for files that are
not excluded, but this shouldn't be a problem because gyp
itself already warns in this case.

See this page for more information:

  <http://connect.microsoft.com/VisualStudio/feedback/details/797460/incorrect-warning-msb8027-reported-for-files-excluded-from-build>

Patch from stk.haller@googlemail.com.

R=scottmg@chromium.org

Review URL: https://codereview.chromium.org/173843004

git-svn-id: http://gyp.googlecode.com/svn/trunk@1856 78cadc50-ecff-11dd-a971-7dbc132099af