Remove pylintrc and enable pylint checks for the Xcode generator.
This copy of pylintrc is never used, and its presence is confusing. The
canned pylint check uses the pylintrc in depot_tools. Project-specific
blacklisted and whitelisted files and disabled warnings are enumerated
in PRESUBMIT.py.

xcode.py is pylint-clean with the current pylint configuration, so it no
longer needs to be excluded from pylint checks.

TEST=presubmit
R=thakis@chromium.org

Review URL: https://codereview.chromium.org/263213006

git-svn-id: http://gyp.googlecode.com/svn/trunk@1917 78cadc50-ecff-11dd-a971-7dbc132099af