android: set GYP_VAR_PREFIX as a local variable in actions/rules.
A few actions/rules need to be able to call the intermediates-dir-for
function from the Android build system in order to discover the location
of another target's intermediates. To do this correctly in a multilib
build we need to pass GYP_VAR_PREFIX to pick the right intermediate
location, but since the variable is not expanded until the command is
executed we need to set a rule-specific local variable for this purpose,
as otherwise it will use whichever value of GYP_VAR_PREFIX was set last
instead of the value for the current target.

This is the same pattern used for other target-specific variables such
as LOCAL_PATH.

BUG=
R=thakis@chromium.org

Review URL: https://codereview.chromium.org/273683003

git-svn-id: http://gyp.googlecode.com/svn/trunk@1918 78cadc50-ecff-11dd-a971-7dbc132099af