android: don't try to handle suppress_wildcard.
When deciding whether to include a target in the generated makefiles or
not, don't consider the suppress_wildcard flag as this just means that
any target which depends directly on the target in question will crash
gyp. The exact contents of the "all" target on the android backend are
not really relevant, so this is the simplest fix.

BUG=chromium:435772
R=thakis@chromium.org

Review URL: https://codereview.chromium.org/758943002

git-svn-id: http://gyp.googlecode.com/svn/trunk@2010 78cadc50-ecff-11dd-a971-7dbc132099af