Add support for generating an Eclipse .classpath file (bug fix)
In r2015, support was added for generating a .classpath file in
the Eclipse generator.  However, it contained a bug that would
add an incorrect directory to the classpath in some circumstances,
if the root folder or any of its parents were named 'src' or
'java'.  This is due to the heuristic nature of determining Java
package root locations.

This commit resolves the bug by preventing the heuristic from
using the root location, or traversing to any of its ancestors.

BUG=gyp:472
Review URL: https://codereview.chromium.org/798133002/

Patch from Kevin McCormick <mckev@amazon.com>!


git-svn-id: http://gyp.googlecode.com/svn/trunk@2021 78cadc50-ecff-11dd-a971-7dbc132099af